+++ src/modules/controls/orientation/components/poi/PoiOrientation.vue
@@ -171,7 +171,7 @@ export default {
             if (style) {
                 const featureStyle = style.createStyle(feat, false);
 
-                if (featureStyle?.getImage() instanceof Icon) {
+                if (featureStyle?.getImage?.() instanceof Icon) {
                     imagePath = featureStyle.getImage()?.getSrc() ? featureStyle.getImage()?.getSrc() : "";
                 }
                 else {
@@ -183,7 +183,7 @@ export default {
                             imagePath = this.createLineSVG(legendInfo.styleObject);
                         }
                         else if (legendInfo.geometryType === "Polygon") {
-                            imagePath = this.createPolygonSVG(legendInfo.styleObject);
+                            imagePath = this.createPolygonGraphic(legendInfo.styleObject);
                         }
                     });
                 }
@@ -271,17 +271,22 @@ export default {
         },
 
         /**
-         * Creating the polygon svg
+         * Creating the polygon graphic
          * @param  {ol/style} style ol style
-         * @return {string} SVG
+         * @return {string} SVG or data URL
          */
-        createPolygonSVG (style) {
+        createPolygonGraphic (style) {
             let svg = "";
-            const fillColor = style.returnColor(style.get("polygonFillColor"), "hex"),
+            const fillColor = style.returnColor(style.get("polygonFillColor") || "black", "hex"),
                 strokeColor = style.returnColor(style.get("polygonStrokeColor"), "hex"),
                 strokeWidth = parseInt(style.get("polygonStrokeWidth"), 10),
-                fillOpacity = style.get("polygonFillColor")[3].toString() || 0,
-                strokeOpacity = style.get("polygonStrokeColor")[3].toString() || 0;
+                fillOpacity = style.get("polygonFillColor")?.[3]?.toString() || 0,
+                strokeOpacity = style.get("polygonStrokeColor")[3].toString() || 0,
+                fillHatch = style.get("polygonFillHatch");
+
+            if (fillHatch) {
+                return style.getPolygonFillHatchLegendDataUrl();
+            }
 
             svg += "<svg height='35' width='35'>";
             svg += "<polygon points='5,5 30,5 30,30 5,30' style='fill:";